Fix 100-print_python_list_info.c 98

+++ 0x03-python-data_structures/100-print_python_list_info.c
@@ -13,13 +13,13 @@ void print_python_list_info(PyObject *p)
 
 	len = PyList_Size(p);
 
-	printf("[*] Size of the Python List = %d\n", len);
-	printf("[*] Allocated = %d\n", (int)((PyListObject *)p)->allocated);
+	printf("[*] Size of the Python List = %ld\n", len);
+	printf("[*] Allocated = %ld\n", (signed long)((PyListObject *)p)->allocated);
 
 	for (i = 0; i < len; i++)
 	{
 		type = Py_Type(((PyListObject *)p)->ob_item[i]->tp_name);
-		printf("Element %d: %s\n", i, type);
+		printf("Element %ld: %s\n", i, type);
 	}
 }
